+++ src/InnerList.php
@@ -74,11 +74,6 @@ final class InnerList implements Countable, IteratorAggregate, StructuredField, 
         return $this->parameters;
     }
 
-    public function parameter(string $key): Item|Token|ByteSequence|float|int|bool|string
-    {
-        return $this->parameters->get($key)->value();
-    }
-
     /**
      * @param Parameters|iterable<array-key, Item|Token|ByteSequence|float|int|bool|string> $parameters
      */